requirements.txt
# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Data fetching
yfinance>=0.2.0
requests>=2.31.0

# Orchestration
prefect==2.20.18

# Analysis
scipy>=1.11.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Advanced ML
xgboost>=2.0.0
lightgbm>=4.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
shap>=0.42.0

# NLP / RAG
transformers>=4.30.0
faiss-cpu>=1.7.0

# Visualization (dashboard)
streamlit>=1.28.0
plotly>=5.17.0
folium>=0.14.0
streamlit-folium>=0.15.0

# Visualization (notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
python-dotenv>=1.0.0
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
ipykernel>=6.0.0
jupyter>=1.0.0
